+++ src/extractAttributes.py
@@ -9,9 +9,9 @@
 
 from __future__ import annotations
 import argparse, json, csv, os
-from copy import deepcopy
-from dataclasses import dataclass, asdict
-from typing import Any, List, Dict, Set, Union
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Dict, Set, Union
 from docx import Document
 from docx.table import Table
 import docx.opc.exceptions
@@ -30,6 +30,7 @@ class AttributeTable:
 	attribute:int
 	shortname:int
 	occursIn:int
+	filename:str
 	category:str
 
 @dataclass
@@ -46,11 +47,11 @@ class Attribute:
 	def asDict(self) -> dict:
 		"""	Return this dataclass as a dictionary.
 		"""
-		return 	{	'shortname' :	self.shortname,
-					'attribute' :	self.attribute,
-					'occursIn' :	[ v for v in self.occursIn ],
-					'categories' :	[ v for v in self.categories ],
-					'documents' :	[ v for v in self.documents ]
+		return 	{	'shortname'	:	self.shortname,
+					'attribute'	:	self.attribute,
+					'occursIn'	:	sorted([ v for v in self.occursIn ]),
+					'categories':	sorted([ v for v in self.categories ]),
+					'documents'	:	sorted([ v for v in self.documents ])
 				}
 
 Attributes = Dict[str, Attribute]
@@ -66,29 +67,38 @@ console = Console()
 attributeTables:list[AttributeTable] = [
 
 	# TS-0004
-	AttributeTable(headers=['Parameter Name', 'XSD long name', 'Occurs in', 'Short Name'],	attribute=1, shortname=3, occursIn=2,  category='Primitive Parameters'),
-	AttributeTable(headers=['Root Element Name', 'Occurs in', 'Short Name'], 				attribute=0, shortname=2, occursIn=1,  category='Primitive Root Elements'),
-	AttributeTable(headers=['Attribute Name', 'Occurs in', 'Short Name'], 					attribute=0, shortname=2, occursIn=1,  category='Resource Attributes'),
-	AttributeTable(headers=['Resource Type Name', 'Short Name'], 							attribute=0, shortname=1, occursIn=-1, category='Resource Types'),
-	AttributeTable(headers=['Member Name', 'Occurs in', 'Short Name'],						attribute=0, shortname=2, occursIn=1,  category='Complex Data Types'),
-	AttributeTable(headers=['Member Name', 'Short Name'],									attribute=0, shortname=1, occursIn=-1, category='Trigger Payload Fields'),
-
-	# TS-0023
-	AttributeTable(headers=['Argument Name', 'Occurs in', 'Short Name'],					attribute=0, shortname=2, occursIn=1,  category='Action Arguments'),
-	AttributeTable(headers=['Returned Value Name', 'Occurs in', 'Short Name'],				attribute=0, shortname=2, occursIn=1,  category='Action Return Values'),
+	AttributeTable(headers=['Parameter Name', 'XSD long name', 'Occurs in', 'Short Name'],	attribute=1, shortname=3, occursIn=2,  filename='ts-0004', category='Primitive Parameters'),
+	AttributeTable(headers=['Root Element Name', 'Occurs in', 'Short Name'], 				attribute=0, shortname=2, occursIn=1,  filename='ts-0004', category='Primitive Root Elements'),
+	AttributeTable(headers=['Attribute Name', 'Occurs in', 'Short Name'], 					attribute=0, shortname=2, occursIn=1,  filename='ts-0004', category='Resource Attributes'),
+	AttributeTable(headers=['Resource Type Name', 'Short Name'], 							attribute=0, shortname=1, occursIn=-1, filename='ts-0004', category='Resource Types'),
+	AttributeTable(headers=['Member Name', 'Occurs in', 'Short Name'],						attribute=0, shortname=2, occursIn=1,  filename='ts-0004', category='Complex Data Types'),
+	AttributeTable(headers=['Member Name', 'Short Name'],									attribute=0, shortname=1, occursIn=-1, filename='ts-0004', category='Trigger Payload Fields'),
 
 	# TS-0022
-	AttributeTable(headers=['Attribute Name', 'Occurs in', 'Short Name', 'Notes'], 			attribute=0, shortname=2, occursIn=1,  category='Common and Field Device Configuration'),
-	AttributeTable(headers=['Member Name', 'Occurs in', 'Short Name', 'Notes'],				attribute=0, shortname=2, occursIn=1,  category='Complex Data Types'),
+	AttributeTable(headers=['Attribute Name', 'Occurs in', 'Short Name', 'Notes'], 			attribute=0, shortname=2, occursIn=1,  filename='ts-0022', category='Common and Field Device Configuration'),
+	AttributeTable(headers=['Member Name', 'Occurs in', 'Short Name', 'Notes'],				attribute=0, shortname=2, occursIn=1,  filename='ts-0022', category='Complex Data Types'),
+	AttributeTable(headers=['ResourceType Name', 'Short Name'], 							attribute=0, shortname=1, occursIn=-1, filename='ts-0022', category='Resource Types'),		# Circumventing a typo in TS-0022 
+
+	# TS-0023
+	AttributeTable(headers=['Resource Type Name', 'Short Name'], 							attribute=0, shortname=1, occursIn=-1, filename='ts-0023', category='Specialization type short names'),
+	AttributeTable(headers=['Attribute Name', 'Occurs in', 'Short Name'], 					attribute=0, shortname=2, occursIn=1,  filename='ts-0004', category='Resource attribute short names'),
+	AttributeTable(headers=['Argument Name', 'Occurs in', 'Short Name'],					attribute=0, shortname=2, occursIn=1,  filename='ts-0023', category='Resource attribute short names'),
+
+	# TS-0032
+	AttributeTable(headers=['Attribute Name', 'Short Name'], 								attribute=0, shortname=1, occursIn=-1, filename='ts-0032', category='Security-specific Resource Type Short Names'),
+	AttributeTable(headers=['Attribute Name', 'Occurs in', 'Short Name', 'Notes'], 			attribute=0, shortname=2, occursIn=1,  filename='ts-0032', category='Security-specific oneM2M Attribute Short Names'),
+	AttributeTable(headers=['Member Name', 'Occurs in', 'Short Name', 'Notes'], 			attribute=0, shortname=2, occursIn=1,  filename='ts-0032', category='Security-specific oneM2M Complex data type member short names'),
+
 ]
 
 
 
-def findAttributeTable(table:Table) -> Union[AttributeTable, None]:
+def findAttributeTable(table:Table, filename:str) -> Union[AttributeTable, None]:
 	"""	Search and return a fitting AttributeTable for the given document table.
 		Return `None` if no fitting entry could be found.
 	"""
 	try:
+		fn = filename.lower()
 		row0 = table.rows[0]
 		for snt in attributeTables:
 			if len(snt.headers) != len(row0.cells):
@@ -98,6 +108,7 @@ def findAttributeTable(table:Table) -> Union[AttributeTable, None]:
 			while isMatch and idx < len(snt.headers):
 				isMatch = row0.cells[idx].text == snt.headers[idx]
 				idx += 1
+			isMatch = isMatch and fn.startswith(snt.filename)
 			if isMatch:
 				return snt
 	except:
@@ -105,7 +116,7 @@ def findAttributeTable(table:Table) -> Union[AttributeTable, None]:
 	return None
 
 
-def processDocuments(documents:list[str], outFilename:str, csvOut:bool) -> Attributes|None:
+def processDocuments(documents:list[str], outDirectory:str, csvOut:bool) -> Attributes|None:
 
 	docs 							= {}
 	ptasks 							= {}
@@ -116,6 +127,12 @@ def processDocuments(documents:list[str], outFilename:str, csvOut:bool) -> Attri
 					BarColumn(),
 					TextColumn('[progress.percentage]{task.percentage:>3.0f}%'),
 					speed_estimate_period=2.0) as progress:
+		
+		def stopProgress(msg:str='') -> None:
+			progress.stop()
+			progress.remove_task(readTask)
+			console.print(msg)
+
 
 		# Preparing tasks for progress
 		readTask 	= progress.add_task(f'Reading document{"s" if len(documents)>1 else ""} ...', total=len(documents))
@@ -125,19 +142,21 @@ def processDocuments(documents:list[str], outFilename:str, csvOut:bool) -> Attri
 		#
 
 		for d in documents:
+			if not (dp := Path(d)).exists():
+				stopProgress(f'[red]Input document "{d}" does not esist')
+				return None
+			if not dp.is_file():
+				stopProgress(f'[red]Input document "{d}" is not a file')
+				return None
 			try:
 				docs[d] = Document(d)
 				ptasks[d] = progress.add_task(f'Processing {d} ...', total=1000)
 				progress.update(readTask, advance=1)
 			except docx.opc.exceptions.PackageNotFoundError as e:
-				progress.stop()
-				progress.remove_task(readTask)
-				console.print(f'[red]Input document "{d}" is not a .docx file')
+				stopProgress(f'[red]Input document "{d}" is not a .docx file')
 				return None
 			except Exception as e:
-				progress.stop()
-				progress.remove_task(readTask)
-				console.print(f'[red]Error reading file "{d}"')
+				stopProgress(f'[red]Error reading file "{d}"')
 				console.print_exception()
 				return None
 		
@@ -156,7 +175,7 @@ def processDocuments(documents:list[str], outFilename:str, csvOut:bool) -> Attri
 			progress.update(processTask, total=len(doc.tables))
 			for table in doc.tables:
 				progress.update(processTask, advance=1)
-				if (snt := findAttributeTable(table)) is None:
+				if (snt := findAttributeTable(table, docName)) is None:
 					continue
 				headersLen = len(snt.headers)
 				for r in table.rows[1:]:
@@ -165,9 +184,9 @@ def processDocuments(documents:list[str], outFilename:str, csvOut:bool) -> Attri
 						continue
 
 					# Extract names and do a bit of transformations
-					attribute = unidecode(cells[snt.attribute].text).strip()
-					shortname = unidecode(cells[snt.shortname].text.replace('*', '').lower())
-					occursIn = map(str.strip, unidecode(cells[snt.occursIn].text).split(',')) if snt.occursIn > -1 else ['n/a']	# Split and strip 'occurs in' entries
+					attributeName	= unidecode(cells[snt.attribute].text).strip()
+					shortname 		= unidecode(cells[snt.shortname].text.replace('*', '').lower())
+					occursIn 		= map(str.strip, unidecode(cells[snt.occursIn].text).split(',')) if snt.occursIn > -1 else ['n/a']	# Split and strip 'occurs in' entries
 					
 					# Don't process empty shortnames
 					if not shortname:	
@@ -183,7 +202,7 @@ def processDocuments(documents:list[str], outFilename:str, csvOut:bool) -> Attri
 						entry.occurences += 1
 					else:
 						entry = Attribute(	shortname=shortname,
-											attribute=attribute,
+											attribute=attributeName,
 											occurences=1,
 											occursIn=set([ v for v in occursIn ]),
 											categories=set([ snt.category ]),
@@ -201,7 +220,7 @@ def processDocuments(documents:list[str], outFilename:str, csvOut:bool) -> Attri
 
 		# count duplicates
 		countDuplicates = 0
-		for shortname,attribute in attributes.items():
+		for shortname, attribute in attributes.items():
 			countDuplicates += 1 if attribute.occurences > 1 else 0
 		progress.update(checkTask, advance=1)
 
@@ -211,9 +230,7 @@ def processDocuments(documents:list[str], outFilename:str, csvOut:bool) -> Attri
 
 		# Write JSON output to a file
 		progress.update(writeTask, advance=1)
-		outFilename += '.json' if not outFilename.endswith('.json') else ''	# fix output filename's extension
-
-		with open(outFilename, 'w') as jsonFile:
+		with open(f'{outDirectory}{os.sep}attributes.json', 'w') as jsonFile:
 			json.dump([ v.asDict() for v in attributes.values()], jsonFile, indent=4)
 
 		# Write output to CSV files
@@ -221,11 +238,13 @@ def processDocuments(documents:list[str], outFilename:str, csvOut:bool) -> Attri
 			for docName, doc in docs.items():			# Individually for each input file
 				progress.update(writeTask, advance=1)
 				# write a sorted list of attribute / shortnames to a csv file
-				with open(f'{os.path.dirname(outFilename)}{os.sep}{docName.rsplit(".", 1)[0] + ".csv"}', 'w') as csvFile:
-					csv.writer(csvFile).writerows(	
+				with open(f'{outDirectory}{os.sep}{docName.rsplit(".", 1)[0] + ".csv"}', 'w') as csvFile:
+					writer = csv.writer(csvFile)
+					writer.writerow(['Attribute', 'Short Name'])
+					writer.writerows(	
 						sorted(
 							[ [attr.attribute, attr.shortname] for attr in attributes.values() if docName in attr.documents ],
-							key=lambda x: x[0].lower() ))
+							key=lambda x: x[0].lower() ))	# type: ignore [index]
 
 		progress.update(writeTask, advance=1)
 
@@ -249,7 +268,8 @@ def printAttributeTables(attributes:Attributes, duplicatesOnly:bool=True) -> Non
 	table.add_column('shortname', no_wrap=True, min_width=6)
 	table.add_column('category', no_wrap=False)
 	table.add_column('document(s)', no_wrap=False)
-	for sn, attribute in attributes.items():
+	for sn in sorted((attributes.keys())):
+		attribute = attributes[sn]
 		if attribute.occurences > 1:
 			table.add_row(attribute.attribute, sn, ', '.join(attribute.categories), f'[red]{", ".join(attribute.documents)}')
 		elif not duplicatesOnly:
@@ -257,12 +277,27 @@ def printAttributeTables(attributes:Attributes, duplicatesOnly:bool=True) -> Non
 	console.print(table)
 
 
+def printAttributeCsv(attributes:Attributes, duplicatesOnly:bool=True, outDirectory:str=None) -> None:
+	"""	Print the found attributes to a CSV file. Optionally print only duplicate entries.
+	"""
+	# Write attributes also to a csv file
+	with open(f'{outDirectory}{os.sep}{"attributes" if not duplicatesOnly else "duplicates"}.csv', 'w') as csvFile:
+		writer = csv.writer(csvFile)
+		writer.writerow(['Attribute', 'Short Name', 'Categories', 'Documents'])
+		for sn in sorted((attributes.keys())):
+			attribute = attributes[sn]
+			if attribute.occurences > 1:
+				writer.writerow([attribute.attribute, sn, ','.join(attribute.categories), ','.join(attribute.documents)])
+			elif not duplicatesOnly:
+				writer.writerow([attribute.attribute, sn, ','.join(attribute.categories), ','.join(attribute.documents)])
+
+
 
 if __name__ == '__main__':
 
 	# Parse command line arguments
 	parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-	parser.add_argument('--outfile', '-o', action='store', dest='outFilename', default='attributes.json', help='specify output filename')
+	parser.add_argument('--outdir', '-o', action='store', dest='outDirectory', default='out', metavar='<output directory>',  help='specify output directory')
 	parser.add_argument('--csv', '-c', action='store_true', dest='csvOut', default=False, help='additionally generate shortname csv files')
 	
 	listArgs = parser.add_mutually_exclusive_group()
@@ -273,7 +308,10 @@ if __name__ == '__main__':
 	args = parser.parse_args()
 
 	# Process documents and print output
-	if (attributes := processDocuments(sorted(args.document), args.outFilename, args.csvOut)) is None:
+	os.makedirs(args.outDirectory, exist_ok=True)
+	if (attributes := processDocuments(sorted(args.document), args.outDirectory, args.csvOut)) is None:
 		exit(1)
 	if args.list or args.listDuplicates:
 		printAttributeTables(attributes, args.listDuplicates)
+		if args.csvOut:
+			printAttributeCsv(attributes, args.listDuplicates, args.outDirectory)